tests: fix crc32_test.v

+++ vlib/hash/crc32/crc32.v
@@ -8,9 +8,9 @@ module crc32
 
 // polynomials
 pub const (
-	ieee       = 0xedb88320
-	castagnoli = 0x82f63b78
-	koopman    = 0xeb31d82e
+	ieee       = u32(0xedb88320)
+	castagnoli = u32(0x82f63b78)
+	koopman    = u32(0xeb31d82e)
 )
 
 // The size of a CRC-32 checksum in bytes.
